+++ src/main/java/com/gugucon/shopping/order/service/OrderService.java
@@ -1,27 +1,32 @@
 package com.gugucon.shopping.order.service;
 
+import com.gugucon.shopping.auth.dto.MemberPrincipal;
 import com.gugucon.shopping.common.dto.response.PagedResponse;
 import com.gugucon.shopping.common.exception.ErrorCode;
 import com.gugucon.shopping.common.exception.ShoppingException;
 import com.gugucon.shopping.item.domain.entity.CartItem;
 import com.gugucon.shopping.item.domain.entity.Product;
 import com.gugucon.shopping.item.repository.CartItemRepository;
+import com.gugucon.shopping.item.repository.OrderStatRepository;
 import com.gugucon.shopping.item.repository.ProductRepository;
+import com.gugucon.shopping.member.domain.vo.BirthYearRange;
 import com.gugucon.shopping.order.domain.PayType;
 import com.gugucon.shopping.order.domain.entity.Order;
+import com.gugucon.shopping.order.domain.entity.OrderItem;
 import com.gugucon.shopping.order.dto.request.OrderPayRequest;
 import com.gugucon.shopping.order.dto.response.OrderDetailResponse;
 import com.gugucon.shopping.order.dto.response.OrderHistoryResponse;
 import com.gugucon.shopping.order.dto.response.OrderPayResponse;
 import com.gugucon.shopping.order.dto.response.OrderResponse;
 import com.gugucon.shopping.order.repository.OrderRepository;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
 
 @Service
 @Transactional(readOnly = true)
@@ -31,6 +36,7 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final CartItemRepository cartItemRepository;
     private final ProductRepository productRepository;
+    private final OrderStatRepository orderStatRepository;
 
     @Transactional
     public OrderResponse order(final Long memberId) {
@@ -41,6 +47,13 @@ public class OrderService {
 
         final Order order = Order.from(memberId, cartItems);
         return OrderResponse.from(orderRepository.save(order));
+    }
+
+    @Transactional
+    public void complete(final Order order, final MemberPrincipal principal) {
+        order.getOrderItems().forEach(orderItem -> updateOrderStatBy(principal, orderItem));
+        cartItemRepository.deleteAllByMemberId(principal.getId());
+        order.completePay();
     }
 
     public OrderDetailResponse getOrderDetail(final Long orderId, final Long memberId) {
@@ -56,6 +69,11 @@ public class OrderService {
                                                                               pageable);
 
         return convertToPage(orders);
+    }
+
+    public Order findOrderBy(final Long orderId, final Long memberId) {
+        return orderRepository.findByIdAndMemberId(orderId, memberId)
+                .orElseThrow(() -> new ShoppingException(ErrorCode.INVALID_ORDER));
     }
 
     private PagedResponse<OrderHistoryResponse> convertToPage(final Page<Order> orders) {
@@ -98,5 +116,12 @@ public class OrderService {
             product.validateStockIsNotLessThan(orderItem.getQuantity());
             product.decreaseStockBy(orderItem.getQuantity());
         });
+    }
+
+    private void updateOrderStatBy(final MemberPrincipal principal, final OrderItem orderItem) {
+        orderStatRepository.updateOrderStatByCount(orderItem.getQuantity().getValue(),
+                                                   orderItem.getProductId(),
+                                                   BirthYearRange.from(principal.getBirthDate()),
+                                                   principal.getGender());
     }
 }
