+++ src/main/java/com/gugucon/shopping/pay/service/PayService.java
@@ -4,12 +4,8 @@ import com.gugucon.shopping.auth.dto.MemberPrincipal;
 import com.gugucon.shopping.common.domain.vo.Money;
 import com.gugucon.shopping.common.exception.ErrorCode;
 import com.gugucon.shopping.common.exception.ShoppingException;
-import com.gugucon.shopping.item.repository.CartItemRepository;
-import com.gugucon.shopping.item.repository.OrderStatRepository;
-import com.gugucon.shopping.member.domain.vo.BirthYearRange;
 import com.gugucon.shopping.order.domain.entity.Order;
-import com.gugucon.shopping.order.domain.entity.OrderItem;
-import com.gugucon.shopping.order.repository.OrderRepository;
+import com.gugucon.shopping.order.service.OrderService;
 import com.gugucon.shopping.pay.config.TossPayConfiguration;
 import com.gugucon.shopping.pay.domain.Pay;
 import com.gugucon.shopping.pay.dto.request.PointPayRequest;
@@ -32,24 +28,22 @@ import org.springframework.transaction.annotation.Transactional;
 public class PayService {
 
     private final PayRepository payRepository;
-    private final OrderRepository orderRepository;
-    private final CartItemRepository cartItemRepository;
     private final PointRepository pointRepository;
     private final TossPayProvider tossPayProvider;
     private final TossPayConfiguration tossPayConfiguration;
-    private final OrderStatRepository orderStatRepository;
+    private final OrderService orderService;
 
     @Transactional
     public PayResponse payByPoint(final PointPayRequest pointPayRequest, final MemberPrincipal principal) {
         final Long orderId = pointPayRequest.getOrderId();
         final Long memberId = principal.getId();
-        final Order order = findOrderBy(orderId, memberId);
+        final Order order = orderService.findOrderBy(orderId, memberId);
 
         final Point point = findPointBy(memberId);
         point.use(order.calculateTotalPrice());
-        order.getOrderItems().forEach(orderItem -> updateOrderStatBy(principal, orderItem));
+        orderService.complete(order, principal);
 
-        return completePay(memberId, order);
+        return PayResponse.from(payRepository.save(Pay.from(order)));
     }
 
     private Point findPointBy(final Long memberId) {
@@ -58,7 +52,7 @@ public class PayService {
     }
 
     public TossPayInfoResponse getTossInfo(final Long orderId, final Long memberId) {
-        final Order order = findOrderBy(orderId, memberId);
+        final Order order = orderService.findOrderBy(orderId, memberId);
 
         return TossPayInfoResponse.from(tossPayProvider.encodeOrderId(order.getId(), order.createOrderName()),
                                         order,
@@ -71,31 +65,14 @@ public class PayService {
     public PayResponse payByToss(final TossPayRequest tossPayRequest, final MemberPrincipal principal) {
         final Long orderId = tossPayProvider.decodeOrderId(tossPayRequest.getOrderId());
         final Long memberId = principal.getId();
-        final Order order = findOrderBy(orderId, memberId);
+        final Order order = orderService.findOrderBy(orderId, memberId);
 
         order.validateMoney(Money.from(tossPayRequest.getAmount()));
-        order.getOrderItems().forEach(orderItem -> updateOrderStatBy(principal, orderItem));
+        orderService.complete(order, principal);
+        final Pay pay = payRepository.save(Pay.from(order));
         tossPayProvider.validatePayment(tossPayRequest);
 
-        return completePay(memberId, order);
-    }
-
-    private PayResponse completePay(final Long memberId, final Order order) {
-        cartItemRepository.deleteAllByMemberId(memberId);
-        order.completePay();
-        return PayResponse.from(payRepository.save(Pay.from(order)));
-    }
-
-    private void updateOrderStatBy(final MemberPrincipal principal, final OrderItem orderItem) {
-        orderStatRepository.updateOrderStatByCount(orderItem.getQuantity().getValue(),
-                                                   orderItem.getProductId(),
-                                                   BirthYearRange.from(principal.getBirthDate()),
-                                                   principal.getGender());
-    }
-
-    private Order findOrderBy(final Long orderId, final Long memberId) {
-        return orderRepository.findByIdAndMemberId(orderId, memberId)
-                .orElseThrow(() -> new ShoppingException(ErrorCode.INVALID_ORDER));
+        return PayResponse.from(pay);
     }
 
     public TossPayFailResponse decodeOrderId(final TossPayFailRequest tossPayFailRequest) {
